+++ src/main/java/ch/ethz/idsc/retina/dev/lidar/LidarRotationProvider.java
@@ -22,7 +22,7 @@ public class LidarRotationProvider implements LidarRayDataListener {
   }
 
   @Override
-  public void timestamp(int usec, byte type) {
+  public void timestamp(int usec, int type) {
     this.usec = usec;
   }
