+++ src/main/java/ch/ethz/idsc/retina/dev/lidar/urg04lx/app/Urg04lxFrame.java
@@ -63,7 +63,7 @@ public class Urg04lxFrame implements Urg04lxEventListener, LidarRayDataListener 
   }
 
   @Override
-  public void timestamp(int usec, byte type) {
+  public void timestamp(int usec, int type) {
     // ---
   }
